refact: função do mapa transferida pra route_controller
