+++ backend/controllers/route_controller.py
@@ -1,6 +1,10 @@
+from backend.domain.route_stop.get_route_stop import GetRouteStop
+from backend.domain.route.get_bus_coords import GetBusCoords
 from backend.domain.route.bus_schedule import BusSchedule
 from backend.domain.bus_stop.get_stop import GetStop
 from backend.domain.route.get_route import GetRoute
+from backend.domain.shape.get_shape import GetShape
+from backend.domain.route.draw_map import DrawMap
 from flask import jsonify, make_response
 
 def return_stop_and_routes(stop_id):
@@ -25,3 +29,10 @@ def return_stop_and_routes(stop_id):
     routes_dict['stop_routes'] = BusSchedule.order_by_arrival_time(routes_dict['stop_routes'])
 
     return make_response(jsonify(routes_dict), 200)
+
+def current_position_map(route_id):
+    polygon = GetShape.get_polygon(route_id)
+    bus_coords = GetBusCoords.get_bus_coords(route_id)
+    bus_stops_coords = GetRouteStop.get_coordinates_stops_in_route(route_id)
+    map = DrawMap.draw_map_route_stops_bus_position(polygon, bus_coords, bus_stops_coords)
+    return make_response(jsonify({'map' : map}), 200)
