+++ backend/domain/route/draw_map_impl.py
@@ -38,6 +38,7 @@ class DrawMapImpl:
                 'center': {"lat": center_lat, "lon": center_lon},
                 'style': "open-street-map",
                 'zoom': 14})
+        fig.show()
 
         return fig.to_html()
